# Performance: string builder speedup for Module.String()