--- a/internal/addrs/module_test.go
+++ b/internal/addrs/module_test.go
@@ -55,3 +55,17 @@ func TestModuleEqual_false(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkModuleStringShort(b *testing.B) {
+	module := Module{"a", "b"}
+	for n := 0; n < b.N; n++ {
+		module.String()
+	}
+}
+
+func BenchmarkModuleStringLong(b *testing.B) {
+	module := Module{"southamerica-brazil-region", "user-regional-desktop", "user-name"}
+	for n := 0; n < b.N; n++ {
+		module.String()
+	}
+}