After the change:

```
terraform/internal/addrs$ go test -bench=Module . -benchmem
BenchmarkModuleStringShort-12           15489140                77.80 ns/op           24 B/op          1 allocs/op          
BenchmarkModuleStringLong-12            10807070               111.8 ns/op            80 B/op          1 allocs/op   
```

before the change:
```
BenchmarkModuleStringShort-12            3290968               384.2 ns/op           120 B/op          3 allocs/op          
BenchmarkModuleStringLong-12             1859038               668.4 ns/op           304 B/op          4 allocs/op  
```

About 5x speedup for time, and only single allocation.

Inspiration for this change is perf profiling on a particularly large config which points to memory churn:
```
  29.34%  terraform  terraform         [.] runtime.scanobject                                                               
  12.31%  terraform  terraform         [.] runtime.findObject                                                               
  11.94%  terraform  terraform         [.] runtime.greyobject                                                               
   8.35%  terraform  terraform         [.] runtime.heapBitsSetType                                                          
   7.25%  terraform  terraform         [.] runtime.mallocgc                                                                 
   5.10%  terraform  terraform         [.] strings.Join                                                                     
   3.12%  terraform  terraform         [.] runtime.memmove                                                                  
   3.02%  terraform  terraform         [.] runtime.gcDrain                                                                  
   2.85%  terraform  terraform         [.] google3/third_party/golang/hashicorp/terraform/addrs/addrs.Module.String         
   2.13%  terraform  terraform         [.] runtime.growslice                                                                
   1.65%  terraform  terraform         [.] google3/third_party/golang/hashicorp/terraform/terraform/terraform.(*nodeExpandMo
   1.02%  terraform  terraform         [.] runtime.sweepone                                                                 
   0.99%  terraform  terraform         [.] runtime.(*sweepLocked).sweep                                                     
   0.77%  terraform  terraform         [.] runtime.(*spanSet).pop                                                           
   0.68%  terraform  terraform         [.] runtime.memclrNoHeapPointers     
```