@dennygursky 2022-06-22
see similar idea in https://github.com/hashicorp/terraform/pull/28246