@apparentlymart 2022-06-23
@alisdair I'm not sure if there's a connection here that prompted working on this or if it was just a coincidence, but FWIW just yesterday I happened to spot a place where we were using `addrs.Module.String` to produce map keys for each module in the configuration, over in #31302.

In that PR I was trying a different strategy for potentially improving it -- effectively, doing the `String` calls all up front at the start and comparing their results instead -- but as I noted over there I ran out of time before I was able to write up a benchmark to see how productive the optimization is, and now I see this PR I guess I should make sure to rebase against this change so I'm measuring against the optimized `String` implementation!